Extend System.mount to support prepending.
System.mount now accepts a fourth argument to prepend the mount to the search path instead of append it. In addition, specifying the fourth argument when the third argument is true will update the path cache instead of triggering a complete reload.

This commit also adds a System.regenerate_scripts method, which regenerates the $RGSS_SCRIPTS array, for dynamically loading patches. This method is a no-op after the preload scripts have finished executing.