Properly handle transparent text with outlines.
RGSS's outline is drawn by blitting a complete set of text four times at different offsets, which affects the final text color and opacity for transparent text. We need to replicate it.

The software blitting this commit introduces actually turned out to be faster than calling TTF_SetFontOutline, which apparently clears the font's glyph cache (which is *not* noted in SDL_ttf.h).