Change how hires text sizing works.
Since the ppem doesn't directly correspond to the size now, it makes more sense to apply the hires multiplier to the ppem instead of the size. This commit's implementation has the side effect of not allowing drawing "lores" text directly to a hires bitmap; whatever size you set will always have the multiplier applied on top of it.